refactor: create user football roles module
