+++ src/services/api/methods/user-football-roles.ts
@@ -1,0 +1,8 @@
+import { UserFootballRoleDto } from '@/modules/user-football-roles/types'
+import { getDataList } from '@/services/api/methods/helpers'
+import { TModuleName } from '@/services/api/modules'
+
+const moduleName: TModuleName = 'user-football-roles'
+
+export const getUserFootballRolesList = () =>
+  getDataList<UserFootballRoleDto>(moduleName)
